#4a — move CostSink + pool registry out of runtime/opencode/
Lifts ``CostSink`` type + ``agents_config_for`` / ``cost_sink_for`` /
register / unregister helpers from ``runtime/opencode/registry.py`` to
``runtime/registry.py``. Vendor-neutral concern — the registries hold
airframe ``AgentsConfig`` + an async sink callable; neither concept
belongs inside the OpenCode HTTP package.

All 15 production importers (agents, actor shells, squadrons) +
1 test conftest switch to ``from maverick.runtime.registry import ...``.
The opencode package's ``__init__`` keeps its re-export so any third
party still grepping the old surface keeps working until the package
itself is deleted.

3534 unit tests pass.
