+++ src/maverick/actors/xoscar/decomposer.py
@@ -31,7 +31,7 @@ from maverick.payloads import (
     SubmitOutlinePayload,
 )
 from maverick.runtime.agent_factory import runtime_for_agent
-from maverick.runtime.opencode import agents_config_for, cost_sink_for
+from maverick.runtime.registry import agents_config_for, cost_sink_for
 
 if TYPE_CHECKING:
     from maverick.executor.config import StepConfig
