+++ src/maverick/actors/xoscar/reviewer.py
@@ -27,7 +27,7 @@ from maverick.agents.reviewer import (
 from maverick.logging import get_logger
 from maverick.payloads import SubmitReviewPayload
 from maverick.runtime.agent_factory import runtime_for_agent
-from maverick.runtime.opencode import agents_config_for, cost_sink_for
+from maverick.runtime.registry import agents_config_for, cost_sink_for
 
 if TYPE_CHECKING:
     from maverick.executor.config import StepConfig
