+++ src/maverick/agents/base.py
@@ -35,7 +35,7 @@ from pydantic import BaseModel, ValidationError
 from maverick.logging import get_logger
 
 if TYPE_CHECKING:
-    from maverick.runtime.opencode import CostSink
+    from maverick.runtime.registry import CostSink
 
 logger = get_logger(__name__)
