+++ src/maverick/agents/briefing/agent.py
@@ -18,7 +18,7 @@ if TYPE_CHECKING:
     from airframe.protocol import AgentRuntime
 
     from maverick.executor.config import StepConfig
-    from maverick.runtime.opencode import CostSink
+    from maverick.runtime.registry import CostSink
 
 BRIEFING_TIMEOUT_SECONDS = 1200
