+++ src/maverick/agents/decomposer.py
@@ -29,7 +29,7 @@ if TYPE_CHECKING:
     from airframe.protocol import AgentRuntime
 
     from maverick.executor.config import StepConfig
-    from maverick.runtime.opencode import CostSink
+    from maverick.runtime.registry import CostSink
 
 DETAIL_TIMEOUT_SECONDS = 1200
 DEFAULT_PROMPT_TIMEOUT_SECONDS = 1800
