+++ src/maverick/runtime/registry.py
@@ -8,7 +8,7 @@ registries live here, each indexed by the pool's external address:
   registers ``squadron.config.agents`` so the actor shells' fallback
   ``_make_agent`` paths can construct via :func:`runtime_for_agent`.
 * Cost sinks — optional; when a workflow registers one (typically a
-  :class:`RunwayStore`-backed appender), every successful mailbox send
+  :class:`RunwayStore`-backed appender), every successful agent send
   flushes a :class:`CostEntry` to it.
 """
 
@@ -33,7 +33,7 @@ __all__ = [
 
 # Async callable that accepts a :class:`runway.models.CostEntry`-shaped
 # dataclass. Typed as ``Any`` here to avoid pulling runway into the
-# runtime package — the mixin imports CostEntry locally and the
+# runtime package — the agent imports CostEntry locally and the
 # workflow registers an appender that closes over a RunwayStore.
 CostSink = Callable[[Any], Awaitable[None]]
 
@@ -43,17 +43,7 @@ _cost_sink_by_pool: dict[str, CostSink] = {}
 
 
 def agents_config_for(pool_address: str) -> AgentsConfig | None:
-    """Return the airframe :class:`AgentsConfig` for ``pool_address``.
-
-    The fallback ``_make_agent`` paths in mailbox actor shells call this
-    when no ``agent=`` was injected by the supervisor: with the config
-    in hand, the actor can call :func:`runtime_for_agent` to materialise
-    an airframe runtime for its role.
-
-    ``None`` means no config was registered — the actor's fallback
-    should raise a clear error rather than silently constructing an
-    untargeted agent.
-    """
+    """Return the airframe :class:`AgentsConfig` for ``pool_address``."""
     return _agents_config_by_pool.get(pool_address)
 
 
@@ -71,12 +61,7 @@ def unregister_agents_config(pool_address: str) -> None:
 
 
 def cost_sink_for(pool_address: str) -> CostSink | None:
-    """Return the cost sink for ``pool_address`` or ``None`` when none set.
-
-    Mailbox actors use this to forward each successful send's
-    :class:`CostEntry` to the workflow's chosen aggregator (typically
-    a :class:`RunwayStore`-backed appender).
-    """
+    """Return the cost sink for ``pool_address`` or ``None`` when none set."""
     return _cost_sink_by_pool.get(pool_address)
