+++ src/maverick/squadron/plan.py
@@ -21,7 +21,7 @@ from maverick.squadron.base import Squadron
 
 if TYPE_CHECKING:
     from maverick.config import MaverickConfig
-    from maverick.runtime.opencode import CostSink
+    from maverick.runtime.registry import CostSink
 
 
 class PlanSquadron(Squadron):
